Updated emojis for three movies
Updated Remember the Titans to change rugby football to American football and focus on friendship building aspects of the movie. Updated Kiki's Delivery Service to more clearly link Kiki and the broom.
Updated Home Alone to add a single child and swapped the see no evil monkey for the iconic screaming face.